Adds serializer
