+++ console/program/src/request/input_id/mod.rs
@@ -14,6 +14,7 @@
 // You should have received a copy of the GNU General Public License
 // along with the snarkVM library. If not, see <https://www.gnu.org/licenses/>.
 
+mod bytes;
 mod serialize;
 mod string;
 
@@ -32,79 +33,4 @@ pub enum InputID<N: Network> {
     Record(Group<N>, Field<N>),
     /// The commitment of the external record input.
     ExternalRecord(Field<N>),
-}
-
-impl<N: Network> FromBytes for InputID<N> {
-    /// Reads the input ID from a buffer.
-    fn read_le<R: Read>(mut reader: R) -> IoResult<Self> {
-        // Read the variant.
-        let variant = u8::read_le(&mut reader)?;
-        // Match the variant.
-        match variant {
-            // Constant input.
-            0 => Ok(Self::Constant(Field::read_le(&mut reader)?)),
-            // Public input.
-            1 => Ok(Self::Public(Field::read_le(&mut reader)?)),
-            // Private input.
-            2 => Ok(Self::Private(Field::read_le(&mut reader)?)),
-            // Record input.
-            3 => {
-                // Read the gamma value.
-                let gamma = Group::read_le(&mut reader)?;
-                // Read the serial number.
-                let serial = Field::read_le(&mut reader)?;
-                // Return the record input.
-                Ok(Self::Record(gamma, serial))
-            }
-            // External record input.
-            4 => Ok(Self::ExternalRecord(Field::read_le(&mut reader)?)),
-            // Invalid input.
-            _ => Err(error("Invalid input ID variant")),
-        }
-    }
-}
-
-impl<N: Network> ToBytes for InputID<N> {
-    /// Writes the input ID to a buffer.
-    fn write_le<W: Write>(&self, mut writer: W) -> IoResult<()> {
-        match self {
-            // Constant input.
-            Self::Constant(value) => {
-                // Write the variant.
-                0u8.write_le(&mut writer)?;
-                // Write the value.
-                value.write_le(&mut writer)
-            }
-            // Public input.
-            Self::Public(value) => {
-                // Write the variant.
-                1u8.write_le(&mut writer)?;
-                // Write the value.
-                value.write_le(&mut writer)
-            }
-            // Private input.
-            Self::Private(value) => {
-                // Write the variant.
-                2u8.write_le(&mut writer)?;
-                // Write the value.
-                value.write_le(&mut writer)
-            }
-            // Record input.
-            Self::Record(gamma, serial) => {
-                // Write the variant.
-                3u8.write_le(&mut writer)?;
-                // Write the gamma value.
-                gamma.write_le(&mut writer)?;
-                // Write the serial number.
-                serial.write_le(&mut writer)
-            }
-            // External record input.
-            Self::ExternalRecord(value) => {
-                // Write the variant.
-                4u8.write_le(&mut writer)?;
-                // Write the value.
-                value.write_le(&mut writer)
-            }
-        }
-    }
 }
